pyproject.toml
[tool.poetry]
name = "python-pdf-memory"
version = "0.1.0"
description = ""
authors = ["Jack Danger <jack.danger@datavant.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psutil = "^6.0.0"
pillow = "^11.0.0"
pypdf2 = "^3.0.1"
reportlab = "^4.2.5"
ipython = "^8.28.0"
pypdf = "^5.0.1"
pymupdf = "^1.24.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
